Moved functions to show neighboring points to utils
